# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Api
 * FounderAutopsy API — the first platform where founders document their own startup shutdowns
 * OpenAPI spec version: 0.1.0
 */
export interface RegisterInput {
  email: string;
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
}

export interface HealthStatus {
  status: string;
}

export interface Startup {
  id: number;
  name: string;
  tagline: string;
  industry: string;
  foundedYear: number;
  closedYear: number;
  story: string;
  whatFailed: string;
  lessonsLearned: string;
  /** @nullable */
  peakMrr?: number | null;
  /** @nullable */
  teamSize?: number | null;
  /** @nullable */
  totalRaised?: number | null;
  /** @nullable */
  aiRootCause?: string | null;
  /** @nullable */
  aiFactors?: string[] | null;
  /** @nullable */
  aiTags?: string[] | null;
  /** @nullable */
  aiVerdict?: string | null;
  /** @nullable */
  submittedBy?: number | null;
  createdAt: string;
}

export interface StartupList {
  startups: Startup[];
  total: number;
}

export interface StartupInput {
  name: string;
  tagline: string;
  industry: string;
  foundedYear: number;
  closedYear: number;
  story: string;
  whatFailed: string;
  lessonsLearned: string;
  /** @nullable */
  peakMrr?: number | null;
  /** @nullable */
  teamSize?: number | null;
  /** @nullable */
  totalRaised?: number | null;
}

export interface InsightsSummary {
  totalPostmortems: number;
  totalIndustries: number;
  topFailureCause: string;
}

export interface FailureReason {
  reason: string;
  count: number;
}

export type FailureReasonList = FailureReason[];

export interface IndustryDistribution {
  industry: string;
  count: number;
}

export type IndustryDistributionList = IndustryDistribution[];

export interface FailuresByYear {
  year: number;
  count: number;
}

export type FailuresByYearList = FailuresByYear[];

export interface ErrorResponse {
  error: string;
}

export type ListStartupsParams = {
/**
 * Search by name or tagline
 */
search?: string;
/**
 * Filter by industry
 */
industry?: string;
/**
 * Filter by AI-identified failure cause tag
 */
failureCause?: string;
/**
 * Filter by closed year
 * @nullable
 */
year?: number | null;
limit?: number;
offset?: number;
};

export type Register200 = {
  token: string;
  user: User;
};

export type Login200 = {
  token: string;
  user: User;
};