# lib/api-zod/src/generated/types/insightsSummary.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Api
 * FounderAutopsy API — the first platform where founders document their own startup shutdowns
 * OpenAPI spec version: 0.1.0
 */

export interface InsightsSummary {
  totalPostmortems: number;
  totalIndustries: number;
  topFailureCause: string;
}