# lib/api-zod/src/generated/types/startup.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Api
 * FounderAutopsy API — the first platform where founders document their own startup shutdowns
 * OpenAPI spec version: 0.1.0
 */

export interface Startup {
  id: number;
  name: string;
  tagline: string;
  industry: string;
  foundedYear: number;
  closedYear: number;
  story: string;
  whatFailed: string;
  lessonsLearned: string;
  /** @nullable */
  peakMrr?: number | null;
  /** @nullable */
  teamSize?: number | null;
  /** @nullable */
  totalRaised?: number | null;
  /** @nullable */
  aiRootCause?: string | null;
  /** @nullable */
  aiFactors?: string[] | null;
  /** @nullable */
  aiTags?: string[] | null;
  /** @nullable */
  aiVerdict?: string | null;
  /** @nullable */
  submittedBy?: number | null;
  createdAt: string;
}