# lib/api-zod/src/generated/types/startupInput.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Api
 * FounderAutopsy API — the first platform where founders document their own startup shutdowns
 * OpenAPI spec version: 0.1.0
 */

export interface StartupInput {
  name: string;
  tagline: string;
  industry: string;
  foundedYear: number;
  closedYear: number;
  story: string;
  whatFailed: string;
  lessonsLearned: string;
  /** @nullable */
  peakMrr?: number | null;
  /** @nullable */
  teamSize?: number | null;
  /** @nullable */
  totalRaised?: number | null;
}